requirements.txt
python-telegram-bot==21.10
httpx==0.28.1
python-dotenv==1.1.0
